requirements.txt
achironet-paynow==1.0.0
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
Django==5.0.3
django-crispy-forms==2.1
django-jazzmin==2.6.0
idna==3.6
psycopg2==2.9.9
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.2.1
